+++ src/ReVitae.Core/Projects/CvProjectApplicationInfo.cs
@@ -1,0 +1,9 @@
+using System.Reflection;
+
+namespace ReVitae.Core.Projects;
+
+public static class CvProjectApplicationInfo
+{
+    public static string Version =>
+        Assembly.GetExecutingAssembly().GetName().Version?.ToString(3) ?? "0.1.0";
+}